Add files via upload
This is my first small web site project This is our group project for collage. This  web site in we use HTML, CSS, ASP and SQL.
In this site it displaying more top places in India. It only get information for choice good place for travel. It not get trip for travel. It only display Indian top places.